pyproject.toml
[project]
name = "issuespotter-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.1",
    "celery>=5.6.2",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "nudenet>=3.4.2",
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "qdrant-client[fastembed]>=1.5.0",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "sqlalchemy>=2.0.46",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.0.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "requires_qdrant: mark test as requiring Qdrant to be running",
    "slow: mark test as slow running",
]

[tool.coverage.run]
source = ["app"]
omit = ["app/__pycache__/*", "tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
